SU(m) irreps projectors

LOShadows.projection_matrices — Method

projection_matrices(m::Int, n::Int; load_from_memory = true, save_in_memory = true, infer_last::Bool = true)

Compute projection matrices for m and all i ≤ n and 0 ≤ k ≤ i.

Serialize computed matrix.

Deserialize data on disk when found.

Note

The implementation supports multithreading with julia -t <num_threads>.

Warning

As of today, the implementation only supports the computation of the projectors involving the totally symmetric irrep of $SU(m)$ and its dual.

LOShadows.linearised_fock_basis_indices — Method

linearised_fock_basis_indices(m::Int, n::Int)

Indexes the linearised Fock basis used for the projection matrices.